Add bezier curve promote_to routines for multiple-degree promotion.
Unlike the existing curve promotion code, promote_to can increase the
order of a bezier curve by one or more orders in a single call.  The
promoted curve is identical to the original curve.

The degree of the final promoted curve is inferred from the size of the
passed-in control point matrix.  Rather than allocate an incrementally
larger matrix at each step (and then calling the incremental promotion
routine), the low-order control points are copied to the target matrix
and then promotion occurs in-place.  Doing so without extra storage
requires that the new control points are calculated from end to start.

Two tests of promote_to are introduced.  The first performs a single-order
promotion and compares that the resulting control points are identically
equal to those from the standard incremental promotion routine.  The
second test promotes the same curve by five orders and then evaluates
the point, derivative, second derivative, and curvature at the first, last,
and an intermediate point.  This is the same test that was used for the
incremental promotion routines.

Some of the tests are failing due to insufficient slack in the floating
point comparisons.  The tolerances were manually increased until the
'float' tests passed.  Double and other data types still have failed tests.