Add bezier curve degree_to_cubic routines to cubically approximate curves.
Calculate cubic 'equivalent' to bezier curve.  If input curve has degree
less than cubic, the curve is promoted to an exactly equivalent cubic
curve.  If the input curve is cubic, the output curve is copied from the
input.  If the input curve is higher order, the output curve is a cubic
bezier curve that will match the input curve in endpoint position and
slopes.  This approximation is used because it is very simple and fast.
This curve may be a terrible approximation of the original curve, but
through successive subdivision, a piecewise cubic approximation of any
curve to any accuracy should be possible.

Unlike the current promotion/demotion routines, this routine will
promote/do-nothing/demote as needed.  In addition, unlike the sophisticated
incremental constrained demotion routine using Eck's algorithm, this
approach directly calculates the control points to achieve a cubic curve
that C1 matches the input curve (no matter the order of the input curve).

The code has three main branches.  Promotion, do-nothing, and demotion.
Tests are added for each of these branches.

Promotion from both linear and quadratic curves are tested by evaluating
the point, derivative, second derivative, and curvature at start, end,
and an intermediate point.

The cubic input do-nothing case is tested with exact comparison of the
output control points.

The demotion case is tested by comparing only the point and derivative at
the start and end points.

Some of the tests are failing due to insufficient slack in the floating
point comparisons.  The tolerances were manually increased until the
'float' tests passed.  Double and other data types still have failed tests.