Add bezier curve equiparametric maximum distance routines.
Calculate upper bound of equiparametric distance between bezier curves.
This routine efficiently calculates an upper bound of the distance between
equal parameter points on two bezier curves.  While not useful for
calculating distances between arbitrary curves, this is just the ticket
for judging the quality of one curve meant to approximate another.

This algorithm requires that both curves have identical order, so the
lower order curve is first promoted to match the higher order curve.
Then, the control points of the curves are differenced, resulting in
the control points of the curve that is the equiparametric difference
between the curves.  Since the control points provide a convex hull of
a curve, they represent the worst-case difference between curves.  The
maximum distance between the origin and a control point of this curve
is an upper bound of the equipmarametric distance between the input
curves.

Four tests of these routines are introduced.  The first test checks for
zero distance between a curve and iself.  The second test checks for zero
distance between a curve and an identical promoted version of itself.  The
third test checks for known distance between a curve and an offset version
of itself.  The final test checks for known distance between a curve and
a reversed-parameter version of itself; the curve was constructed such
that the maximum distance would be between the endpoints.