Add private segment-wise splitting routines to piecewise curve.
The outside-facing interface of piecewise curves restricts access to the
parameter t.  However, internally, algorithms can work on a particular
segment.

This change introduces private split_seg methods to split a given
segment at a specified (segment-wise) parameter.  If no parameter is
specified, the segment is bisected.  A further version is introduced
that returns an iterator to the newly added segment.

The existing split routine is refactored to use split_seg.