Add adaptive to_cubic converter for piecewise curves.
This routine converts each segment of a piecewise bezier curve to a cubic
bezier with the same parameterization.  Each segment is adaptively
bisected as required to ensure the maximum distance between the exact
and cubic curves is less than the specified tolerance.

Numerous tests are introduced.  Degenerate cases (where the order of the
input curve is <= 3) result in an exact cubic with no subdivision.
Single-segment and multi-segment piecewise curves are tested.
Approximation of high order curves is tested.  A cubic curve promoted to
high order returns an exact cubic with no subdivision.