Minimum_distance surface solver tends to converge to a nearby point when constrained
There is apparently a more attractive local minimum near the exact solution
when the minimum distance problem is up against a constraint.  Given a
wide variety of starting points, the solver converges to this local minimum
rather than the exact solution.

This changes the test cases to find that nearby local solution and then
verify that a wide variety of starting points converge to the same point.