Refactor bezier::fit to make minimum_distance based error calc optional.
The optimization based minimum_distance calculation performed at the end
of bezier::fit is quite expensive.  Although it is an important diagnostic,
it may not need to be calculated inside a fast loop of a production
program.

This refactoring makes the case where returned 't' values and the error
estimate are not required significantly faster.